Add files via upload
Write a program that would use nested if /if-else. In this program a variable called temper stores a randomly generated number between 1 and 400. If the generated value is above 100, the program prints “ temperature above boiling point”  and uses a second if to check whether the temper is above 320 degrees; in that case the program prints “temperature above smoke point”. For temper lower than 100 the program prints “ temperature is not very high”. Flowchart is required in this program. Student should also suggest what are some of the weaknesses in this problem