Add files via upload
Write a program that takes integer inputs from a user in a while loop in a variable called num and prints the value of num. The loop terminates if the user types a negative number. The program should also count the number of times the loop iterates. Student should provide a flow chart for this problem