Add files via upload
Write a program that uses a list called scores initialized with 10 values. The program iterates over the list and determines student’s grade for the given score, as done in week3 lesson. Student must use elif construct to determine the grades based on the following criteria: score>80 is grade A, score between 70 and 80 is grade B, score between 60 and 70 is grade C, score between 50 and 60 is grade D and score less than 50 is grade D. Use any loop of you choice. Student must provide a flowchart for this program